+++ packages/editor/web/src/RichContentEditor/Toolbars/SideToolbar/SideToolbar.jsx
@@ -86,7 +86,7 @@ export default class SideToolbar extends Component {
           this.setState({
             position: {
               top: top - parentTop + offset.y,
-              [!isMobile ? 'left' : 'right']: offset.x,
+              ...(isMobile ? { right: offset.x } : { left: offset.x, right: offset.x }),
               transform: `scale(${isMobile ? 0.76 : 1})`, //mobile plus is smaller
               transition: 'transform 0.15s cubic-bezier(.3,1.2,.2,1)',
             },
@@ -97,6 +97,7 @@ export default class SideToolbar extends Component {
           position: {
             top: offset.y,
             left: offset.x,
+            right: offset.x,
             transform: `scale(${isMobile ? 0.76 : 1})`, //mobile plus is smaller
             transition: 'transform 0.15s cubic-bezier(.3,1.2,.2,1)',
             position: 'absolute',